+++ Runtime/Scripts/RendererFeatures/UIBlurSourceMode.cs
@@ -1,0 +1,8 @@
+namespace ParkMinPackages.UGUI.Blur.RendererFeatures
+{
+	public enum UIBlurSourceMode
+	{
+		Camera,
+		Image
+	}
+}